## Description
This PR is for speeding up the lanelet filtering process.
This PR will reduce the number of loops by using the inner R-tree of intersected lanelets.

Using the [TimeKeeper](https://github.com/autowarefoundation/autoware.universe/tree/main/common/autoware_universe_utils), I checked the bottleneck of lanelet filter node. The main bottleneck was the for-loop for the filtering for the object that has no bounding box.

### processing times
```Text
objectCallback: 136.19 [ms]
    ├── getConvexHull: 0.35 [ms]
    ├── getIntersectedLanelets: 1.62 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 2.24 [ms]
    │   └── isObjectOverlapLanelets: 2.24 [ms]
    │       └── non_bbox_filter: 2.24 [ms]
    ├── filterObject: 1.77 [ms]
    │   └── isObjectOverlapLanelets: 1.68 [ms]
    │       └── non_bbox_filter: 1.68 [ms]
    ├── filterObject: 1.60 [ms]
    │   └── isObjectOverlapLanelets: 1.60 [ms]
    │       └── non_bbox_filter: 1.60 [ms]
    ├── filterObject: 0.75 [ms]
    │   └── isObjectOverlapLanelets: 0.74 [ms]
    │       └── non_bbox_filter: 0.74 [ms]
    ├── filterObject: 1.11 [ms]
    │   └── isObjectOverlapLanelets: 1.11 [ms]
    │       └── non_bbox_filter: 1.11 [ms]
    ├── filterObject: 3.11 [ms]
    │   └── isObjectOverlapLanelets: 3.11 [ms]
    │       └── non_bbox_filter: 3.10 [ms]
    ├── filterObject: 1.29 [ms]
    │   └── isObjectOverlapLanelets: 1.29 [ms]
    │       └── non_bbox_filter: 1.28 [ms]
    ├── filterObject: 2.51 [ms]
    │   └── isObjectOverlapLanelets: 2.51 [ms]
    │       └── non_bbox_filter: 2.51 [ms]
    ├── filterObject: 1.55 [ms]
    │   └── isObjectOverlapLanelets: 1.55 [ms]
    │       └── non_bbox_filter: 1.55 [ms]
    ├── filterObject: 1.11 [ms]
    │   └── isObjectOverlapLanelets: 1.10 [ms]
    │       └── non_bbox_filter: 1.10 [ms]
    ├── filterObject: 0.54 [ms]
    │   └── isObjectOverlapLanelets: 0.54 [ms]
    │       └── non_bbox_filter: 0.54 [ms]
    ├── filterObject: 2.34 [ms]
    │   └── isObjectOverlapLanelets: 2.34 [ms]
    │       └── non_bbox_filter: 2.33 [ms]
   	...
```


After this PR
```Text
objectCallback: 12.90 [ms]
    ├── getConvexHull: 0.09 [ms]
    ├── getIntersectedLanelets: 0.32 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.00 [ms]
    ├── filterObject: 0.22 [ms]
    │   └── no bbox filter: pre process: 0.21 [ms]
    │       └── no bbox filter: main loop: 0.21 [ms]
    ├── filterObject: 0.58 [ms]
    │   └── no bbox filter: pre process: 0.57 [ms]
    │       └── no bbox filter: main loop: 0.57 [ms]
    ├── filterObject: 0.19 [ms]
    │   └── no bbox filter: pre process: 0.18 [ms]
    │       └── no bbox filter: main loop: 0.18 [ms]
    ├── filterObject: 0.01 [ms]
    │   └── no bbox filter: pre process: 0.01 [ms]
    │       └── no bbox filter: main loop: 0.00 [ms]
    ├── filterObject: 0.23 [ms]
    │   └── no bbox filter: pre process: 0.22 [ms]
    │       └── no bbox filter: main loop: 0.22 [ms]
    ├── filterObject: 0.20 [ms]
    │   └── no bbox filter: pre process: 0.20 [ms]
    │       └── no bbox filter: main loop: 0.20 [ms]
    ├── filterObject: 0.23 [ms]
    │   └── no bbox filter: pre process: 0.23 [ms]
    │       └── no bbox filter: main loop: 0.22 [ms]
    ├── filterObject: 0.18 [ms]
    │   └── no bbox filter: pre process: 0.17 [ms]
    │       └── no bbox filter: main loop: 0.17 [ms]
    ├── filterObject: 0.20 [ms]
    │   └── no bbox filter: pre process: 0.19 [ms]
    │       └── no bbox filter: main loop: 0.19 [ms]
    ├── filterObject: 0.12 [ms]
    │   └── no bbox filter: pre process: 0.12 [ms]
    │       └── no bbox filter: main loop: 0.12 [ms]
    ├── filterObject: 0.53 [ms]
    │   └── no bbox filter: pre process: 0.53 [ms]
    │       └── no bbox filter: main loop: 0.52 [ms]
    ├── filterObject: 0.17 [ms]
    │   └── no bbox filter: pre process: 0.16 [ms]
    │       └── no bbox filter: main loop: 0.16 [ms]
    ├── filterObject: 0.16 [ms]
    │   └── no bbox filter: pre process: 0.15 [ms]
    │       └── no bbox filter: main loop: 0.15 [ms]
    ├── filterObject: 0.05 [ms]
    │   └── no bbox filter: pre process: 0.04 [ms]
    │       └── no bbox filter: main loop: 0.04 [ms]
    ├── filterObject: 0.00 [ms]
    │   └── no bbox filter: pre process: 0.00 [ms]
    │       └── no bbox filter: main loop: 0.00 [ms]
    ├── filterObject: 0.13 [ms]
    │   └── no bbox filter: pre process: 0.13 [ms]
    │       └── no bbox filter: main loop: 0.13 [ms]
    ...
```

### VS. previous implementations
![compare_processing_time](https://github.com/user-attachments/assets/7beb6418-5bc3-4e56-b090-8af60ab0550d)

original: PR before https://github.com/autowarefoundation/autoware.universe/pull/8109
current: PR after https://github.com/autowarefoundation/autoware.universe/pull/8109
optimized: after this PR


##### mean processing time
original mean: 34.401 ms
current mean: 30.980 ms
optimized best comb. mean: 3.899 ms

(Mean processing time was caluculated with record that has more than 1 ms)

## Related links

**Parent Issue:**

- Link

<!-- ⬇️🟢
**Private Links:**

- [TIER4 internal link](https://tier4.atlassian.net/browse/RT1-7219)
⬆️🟢 -->

## How was this PR tested?
Qualitative test is done by checking and comparing with original implementation of detected and filtered unknown objects.
No quantitative evaluation has done.

## Notes for reviewers

None.

## Interface changes

None.

<!-- ⬇️🔴

### Topic changes

#### Additions and removals

| Change type   | Topic Type      | Topic Name    | Message Type        | Description       |
|:--------------|:----------------|:--------------|:--------------------|:------------------|
| Added/Removed | Pub/Sub/Srv/Cli | `/topic_name` | `std_msgs/String`   | Topic description |

#### Modifications

| Version | Topic Type      | Topic Name        | Message Type        | Description       |
|:--------|:----------------|:------------------|:--------------------|:------------------|
| Old     | Pub/Sub/Srv/Cli | `/old_topic_name` | `sensor_msgs/Image` | Topic description |
| New     | Pub/Sub/Srv/Cli | `/new_topic_name` | `sensor_msgs/Image` | Topic description |

### ROS Parameter Changes

#### Additions and removals

| Change type   | Parameter Name | Type     | Default Value | Description       |
|:--------------|:---------------|:---------|:--------------|:------------------|
| Added/Removed | `param_name`   | `double` | `1.0`         | Param description |

#### Modifications

| Version | Parameter Name   | Type     | Default Value | Description       |
|:--------|:-----------------|:---------|:--------------|:------------------|
| Old     | `old_param_name` | `double` | `1.0`         | Param description |
| New     | `new_param_name` | `double` | `1.0`         | Param description |

🔴⬆️ -->

## Effects on system behavior

None.