@YoshiRi 2024-07-30
> LGTM! Thank you for your contribution.
> 
> I did not confirmed that the Autoware coding guideline allows to use abbreviated namespace. `namespace bg = boost::geometry;`

It seems no problem because some planning code use this expression.